+++ tests/unittestcase.py
@@ -216,7 +216,7 @@ class UnitTestCase(TestCase):
                   encoding="utf-8") as f:
             return responses.add(
                 responses.GET,
-                re.compile(f"{self.test_config.constants.ORDER_HISTORY_URL}?.*"),
+                re.compile(rf"^{re.escape(self.test_config.constants.ORDER_HISTORY_URL)}\?.*"),
                 body=f.read(),
                 status=200,
             )
@@ -226,7 +226,17 @@ class UnitTestCase(TestCase):
                   encoding="utf-8") as f:
             return responses.add(
                 responses.GET,
-                re.compile(f"{self.test_config.constants.ORDER_DETAILS_URL}?.*"),
+                re.compile(rf"^{re.escape(self.test_config.constants.ORDER_DETAILS_URL)}\?.*"),
+                body=f.read(),
+                status=200,
+            )
+
+    def given_any_invoice_exists(self, invoice_html_file):
+        with open(os.path.join(self.RESOURCES_DIR, "invoices", invoice_html_file), "r",
+                  encoding="utf-8") as f:
+            return responses.add(
+                responses.GET,
+                re.compile(rf"^{re.escape(self.test_config.constants.ORDER_INVOICE_URL)}\?.*"),
                 body=f.read(),
                 status=200,
             )